+++ src/main/java/step3/domain/LottoStore.java
@@ -1,8 +1,8 @@
-package step3_4.domain;
+package step3.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import step3_4.utils.NumbersGenerator;
+import step3.utils.NumbersGenerator;
 
 public class LottoStore {
