+++ src/main/java/step3/io/InputView.java
@@ -1,4 +1,4 @@
-package step3_4.io;
+package step3.io;
 
 import java.util.Arrays;
 import java.util.List;
