+++ src/test/java/step3/domain/UniqueNumbersTest.java
@@ -1,4 +1,4 @@
-package step3_4.domain;
+package step3.domain;
 
 import java.util.Arrays;
 import java.util.List;
